+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Added support for splines in *3D Viewer* node.
+- Added built-in attributes for Mesh object.
 
 ### Fixed